Add more VM functions. The VM now seems complete... Next: the plugin system!

+++ src/core/vmrunner.cpp
@@ -127,7 +127,6 @@ void blrScriptVM::vmRunFunction(const char *fncnym)
 		int jmp=0;++pinst;
 		//printf("@offset %d, instId 0x%x\n",cur,inst[cur].id);
 		rr[100].r()=0.01667;
-		//printf("r00: %f, r01: %f\n",rr[0].r(),rr[1].r());
 		switch(inst[cur].id)
 		{
 			case 0x00:
@@ -141,8 +140,6 @@ void blrScriptVM::vmRunFunction(const char *fncnym)
 			break;
 			case 0x02:
 				callFnc(inst[cur].para1.fnc);
-				//callStk.clear();
-				//printf("stubbed call %s\n",inst[cur].para1.fnc);
 			break;
 			case 0x03:
 				fetchData(inst[cur].para1,true)=fetchData(inst[cur].para2);
@@ -277,7 +274,7 @@ void blrScriptVM::vmRunFunction(const char *fncnym)
 		if(!jmp)++cur;
 	}
 }
-void blrScriptVM::vmSetRetValf(float v){rr[102].r()=v;}
+void blrScriptVM::vmSetRetValf(double v){rr[102].r()=v;}
 void blrScriptVM::vmSetRetVald(int v){ir[100].i()=v;}
 void blrScriptVM::vmInit(unsigned int seed)
 {
